Sets up SelectionScreen layers and separates from EditorScreen layers.
