+++ app/src/main/java/com/tunepruner/voixt/editor/editorscreen/repository/EditorScreenRepository.kt
@@ -1,0 +1,8 @@
+package com.tunepruner.voixt.editor.editorscreen.repository
+
+import com.tunepruner.voixt.editor.editorscreen.datasource.VoiceToTextDataSourceImpl
+
+class EditorScreenRepository {
+    private val voiceToTextDataSource = VoiceToTextDataSourceImpl()
+
+}
