+++ app/src/main/java/com/tunepruner/voixt/editor/util/IndexResolver.kt
@@ -1,4 +1,4 @@
-package com.tunepruner.voixt.editor.editorscreen.domain
+package com.tunepruner.voixt.editor.util
 
 /** This class will be in charge of resolving the index that will
  * be needed to insert text at the correct position in a text body. */
